dbg: move kvm vcpu debug type to another file
- this is done in preparation of accomodating support for other hypervisors
which means some of the functionality whould be common.
- moving the debug code to a separate file can enable us to
group common behavior in traits
